feat: remove Payload from thea-pallet code
